Replaced "empty"
This way the code is cleaner, and also more likely to work with PHP 5.4.
Tried also to introduce type checking in controller, but gave up the
idea for the time being becasue the criteria seem to be to strict
for current implementation. Maybe one day it can be done, but does not
seem to be very important.